change how various table compoments are initialized
