+++ src/dom_components/model/ComponentTableBody.js
@@ -5,8 +5,41 @@ module.exports = Component.extend({
   defaults: _.extend({}, Component.prototype.defaults, {
     type: 'tbody',
     tagName: 'tbody',
-    droppable: ['tr']
+    droppable: ['tr'],
+    columns: 1,
+    rows: 1,
   }),
+
+  initialize(o, opt) {
+    Component.prototype.initialize.apply(this, arguments);
+    var components = this.get('components');
+    var rows = this.get('rows');
+    var columns = this.get('columns');
+
+    // Init components if empty
+    if(!components.length){
+      var rowsToAdd = [];
+
+      while(rows--){
+        var columnsToAdd = [];
+        var clm = columns;
+
+        while (clm--) {
+          columnsToAdd.push({
+            type: 'cell',
+            classes: ['cell']
+          });
+        }
+
+        rowsToAdd.push({
+          type: 'row',
+          classes: ['row'],
+          components: columnsToAdd
+        });
+      }
+      components.add(rowsToAdd);
+    }
+  },
 
 },{
